+++ lib/widgets/palabras/palabras.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:moblie/model/main.dart';
 import 'package:moblie/model/palabra.model.dart';
+import 'package:moblie/pages/search-screen.dart';
 import 'package:moblie/widgets/palabras/palabras.card.dart';
 import 'package:scoped_model/scoped_model.dart';
 
@@ -9,7 +10,21 @@ class Palabras extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ScopedModelDescendant<MainModel>(
-      builder: (BuildContext context, Widget child, MainModel model) => _buildPalabrasList(model.allPalabras, model)
+      builder: (BuildContext context, Widget child, MainModel model) {
+        return Scaffold(
+          body: _buildPalabrasList(model.allPalabras, model),
+          floatingActionButton: FloatingActionButton(
+            child: Icon(
+              Icons.search, 
+              color: Colors.black54
+            ),
+
+            onPressed: () => Navigator
+              .of(context)
+              .push(searchPalabras(context, model.allPalabras))
+          ),
+        );
+      }
     );
   }
 
@@ -19,8 +34,10 @@ class Palabras extends StatelessWidget {
 
     if (palabras.length > 0) {
       palabrasCard = ListView.builder(
-        itemBuilder: (BuildContext context, int index) => PalabraCard(palabras[index], model),
         itemCount: palabras.length,
+        itemBuilder: (BuildContext context, int index) {
+          return PalabraCard(palabras[index], model);
+        }
       );
     }